Move BQ macros under a scala 2 specific build path
