+++ scio-google-cloud-platform/src/it/scala-2/com/spotify/scio/bigquery/BigQueryPartitionUtilIT.scala
