+++ scio-google-cloud-platform/src/main/scala-2/com/spotify/scio/bigquery/types/SchemaUtil.scala
