+++ scio-google-cloud-platform/src/main/scala-2/com/spotify/scio/bigquery/validation/OverrideTypeProviderFinder.scala
