+++ scio-google-cloud-platform/src/main/scala/com/spotify/scio/bigquery/syntax/ScioContextSyntax.scala
@@ -18,32 +18,24 @@
 package com.spotify.scio.bigquery.syntax
 
 import com.spotify.scio.ScioContext
-import com.spotify.scio.bigquery.types.BigQueryType.HasAnnotation
 import com.spotify.scio.bigquery.{
   BigQuerySelect,
   BigQueryStorage,
   BigQueryStorageSelect,
-  BigQueryType,
-  BigQueryTyped,
   Query,
-  Source,
   Table,
   TableRow,
   TableRowJsonIO
 }
 import com.spotify.scio.coders.Coder
-import com.spotify.scio.schemas.Schema
 import com.spotify.scio.values._
-import org.apache.beam.sdk.io.gcp.bigquery.SchemaAndRecord
 
-import scala.reflect.ClassTag
-import scala.reflect.runtime.universe._
 import com.spotify.scio.bigquery.BigQueryTypedTable
 import com.spotify.scio.bigquery.BigQueryTypedTable.Format
 import com.spotify.scio.bigquery.coders.tableRowCoder
 
-/** Enhanced version of [[ScioContext]] with BigQuery methods. */
-final class ScioContextOps(private val self: ScioContext) extends AnyVal {
+sealed private[syntax] trait ScioContextBaseOps {
+  private[syntax] val self: ScioContext
 
   /**
    * Get an SCollection for a BigQuery SELECT query.
@@ -118,112 +110,13 @@ final class ScioContextOps(private val self: ScioContext) extends AnyVal {
   def bigQueryStorage(query: Query): SCollection[TableRow] =
     self.read(BigQueryStorageSelect(query))
 
-  def typedBigQuery[T <: HasAnnotation: ClassTag: TypeTag: Coder](): SCollection[T] =
-    typedBigQuery(None)
-
-  def typedBigQuery[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    newSource: Source
-  ): SCollection[T] = typedBigQuery(Option(newSource))
-
-  /** Get a typed SCollection for BigQuery Table or a SELECT query using the Storage API. */
-  def typedBigQuery[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    newSource: Option[Source]
-  ): SCollection[T] = {
-    val bqt = BigQueryType[T]
-    if (bqt.isStorage) {
-      newSource
-        .asInstanceOf[Option[Table]]
-        .map(typedBigQueryStorage(_))
-        .getOrElse(typedBigQueryStorage())
-    } else {
-      self.read(BigQueryTyped.dynamic[T](newSource))
-    }
-  }
-
-  @deprecated("Beam SQL support will be removed in 0.11.0", since = "0.10.1")
-  def typedBigQueryTable[T: Schema: Coder: ClassTag](table: Table): SCollection[T] =
-    self.read(BigQueryTyped.BeamSchema(table))
-
-  @deprecated("Beam SQL support will be removed in 0.11.0", since = "0.10.1")
-  def typedBigQueryTable[T: Schema: Coder: ClassTag](
-    table: Table,
-    parseFn: SchemaAndRecord => T
-  ): SCollection[T] =
-    self.read(BigQueryTyped.BeamSchema(table, parseFn))
-
-  /**
-   * Get a typed SCollection for a BigQuery storage API.
-   *
-   * Note that `T` must be annotated with
-   * [[com.spotify.scio.bigquery.types.BigQueryType.fromSchema BigQueryType.fromStorage]] or
-   * [[com.spotify.scio.bigquery.types.BigQueryType.fromQuery BigQueryType.fromQuery]]
-   */
-  def typedBigQueryStorage[T <: HasAnnotation: ClassTag: TypeTag: Coder](): SCollection[T] = {
-    val bqt = BigQueryType[T]
-    if (bqt.isQuery) {
-      self.read(BigQueryTyped.StorageQuery[T](Query(bqt.query.get)))
-    } else {
-      val table = Table.Spec(bqt.table.get)
-      val rr = bqt.rowRestriction
-      val fields = bqt.selectedFields.getOrElse(BigQueryStorage.ReadParam.DefaultSelectFields)
-      self.read(BigQueryTyped.Storage[T](table, fields, rr))
-    }
-  }
-
-  def typedBigQueryStorage[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    table: Table
-  ): SCollection[T] =
-    self.read(
-      BigQueryTyped.Storage[T](
-        table,
-        BigQueryType[T].selectedFields.getOrElse(BigQueryStorage.ReadParam.DefaultSelectFields),
-        BigQueryType[T].rowRestriction
-      )
-    )
-
-  def typedBigQueryStorage[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    rowRestriction: String
-  ): SCollection[T] = {
-    val bqt = BigQueryType[T]
-    val table = Table.Spec(bqt.table.get)
-    self.read(
-      BigQueryTyped.Storage[T](
-        table,
-        bqt.selectedFields.getOrElse(BigQueryStorage.ReadParam.DefaultSelectFields),
-        Option(rowRestriction)
-      )
-    )
-  }
-
-  def typedBigQueryStorage[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    table: Table,
-    rowRestriction: String
-  ): SCollection[T] =
-    self.read(
-      BigQueryTyped.Storage[T](
-        table,
-        BigQueryType[T].selectedFields.getOrElse(BigQueryStorage.ReadParam.DefaultSelectFields),
-        Option(rowRestriction)
-      )
-    )
-
-  def typedBigQueryStorage[T <: HasAnnotation: ClassTag: TypeTag: Coder](
-    table: Table,
-    selectedFields: List[String],
-    rowRestriction: String
-  ): SCollection[T] =
-    self.read(
-      BigQueryTyped.Storage[T](
-        table,
-        selectedFields,
-        Option(rowRestriction)
-      )
-    )
-
   /** Get an SCollection for a BigQuery TableRow JSON file. */
   def tableRowJsonFile(path: String): SCollection[TableRow] =
     self.read(TableRowJsonIO(path))
 }
+
+/** Enhanced version of [[ScioContext]] with BigQuery methods. */
+final class ScioContextOps(private[syntax] val self: ScioContext) extends ScioContextBaseOps
 
 trait ScioContextSyntax {
   implicit def bigQueryScioContextOps(sc: ScioContext): ScioContextOps = new ScioContextOps(sc)
