+++ scio-google-cloud-platform/src/test/scala-2/com/spotify/scio/bigquery/types/Schemas.scala
