+++ scio-google-cloud-platform/src/test/scala-2/com/spotify/scio/bigquery/validation/BigQueryValidationTest.scala
